--- Heroku_Info
@@ -1,5 +1,0 @@
-username: rjxdtdvfxmxaxb
-Password: 0d5f12199551cc34e54ef58785dcabe23c4138007d3368aaece3265e24ab3210@
-Host: ec2-50-17-189-165.compute-1.amazonaws.com
-Port: 5432
-name: d85uf0r6sr3b7t
