+++ client/src/css/Gallery.css
